apps/ocr/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
paddleocr==2.7.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-dateutil==2.9.0.post0
python-mrz==0.6.2
